+++ Classes/Domain/VariableResolvers/LanguageVariableResolver.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
 
 namespace Brotkrueml\JobRouterBase\Domain\VariableResolvers;
 
-use Brotkrueml\JobRouterBase\Enumeration\FieldTypeEnumeration;
+use Brotkrueml\JobRouterBase\Enumeration\FieldType;
 use Brotkrueml\JobRouterBase\Event\ResolveFinisherVariableEvent;
 use Brotkrueml\JobRouterBase\Exception\VariableResolverException;
 
@@ -74,16 +74,15 @@ final class LanguageVariableResolver
 
     private function checkValidFieldTypes(ResolveFinisherVariableEvent $event): void
     {
-        if ($event->getFieldType() === FieldTypeEnumeration::TEXT) {
+        if ($event->getFieldType() === FieldType::Text) {
             return;
         }
 
         throw new VariableResolverException(
             \sprintf(
-                'The value "%s" contains a variable which can only be used in Text fields ("%d"), type "%d" used',
+                'The value "%s" contains a variable which can only be used in "Text" fields, type "%s" used',
                 $event->getValue(),
-                FieldTypeEnumeration::TEXT,
-                $event->getFieldType()
+                $event->getFieldType()->name
             ),
             1582654966
         );