+++ Tests/Unit/Domain/VariableResolvers/LanguageVariableResolverTest.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
 namespace Brotkrueml\JobRouterBase\Tests\Unit\Domain\VariableResolvers;
 
 use Brotkrueml\JobRouterBase\Domain\VariableResolvers\LanguageVariableResolver;
-use Brotkrueml\JobRouterBase\Enumeration\FieldTypeEnumeration;
+use Brotkrueml\JobRouterBase\Enumeration\FieldType;
 use Brotkrueml\JobRouterBase\Event\ResolveFinisherVariableEvent;
 use Brotkrueml\JobRouterBase\Exception\VariableResolverException;
 use PHPUnit\Framework\MockObject\Stub;
@@ -65,7 +65,7 @@ class LanguageVariableResolverTest extends TestCase
     public function languageVariablesAreResolvedCorrectly(string $value, string $expected): void
     {
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::TEXT,
+            FieldType::Text,
             $value,
             '',
             [],
@@ -141,7 +141,7 @@ class LanguageVariableResolverTest extends TestCase
     public function multipleLanguageVariablesAreResolved(): void
     {
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::TEXT,
+            FieldType::Text,
             '{__language.twoLetterIsoCode} {__language.direction}',
             '',
             [],
@@ -159,7 +159,7 @@ class LanguageVariableResolverTest extends TestCase
     public function onlyLanguageVariablesAreResolved(): void
     {
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::TEXT,
+            FieldType::Text,
             '{__language1.twoLetterIsoCode}',
             '',
             [],
@@ -177,7 +177,7 @@ class LanguageVariableResolverTest extends TestCase
     public function languageKeyThatCannotMatchedIsIgnored(): void
     {
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::TEXT,
+            FieldType::Text,
             '{__language.invalid key}',
             '',
             [],
@@ -196,10 +196,10 @@ class LanguageVariableResolverTest extends TestCase
     {
         $this->expectException(VariableResolverException::class);
         $this->expectExceptionCode(1582654966);
-        $this->expectExceptionMessage('The value "{__language.twoLetterIsoCode}" contains a variable which can only be used in Text fields ("1"), type "2" used');
+        $this->expectExceptionMessage('The value "{__language.twoLetterIsoCode}" contains a variable which can only be used in "Text" fields, type "Integer" used');
 
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::INTEGER,
+            FieldType::Integer,
             '{__language.twoLetterIsoCode}',
             '',
             [],
@@ -221,7 +221,7 @@ class LanguageVariableResolverTest extends TestCase
             ->willReturn(null);
 
         $event = new ResolveFinisherVariableEvent(
-            FieldTypeEnumeration::TEXT,
+            FieldType::Text,
             '{__language.twoLetterIsoCode}',
             '',
             [],